requirements.txt
Flask==2.2.2
Werkzeug==2.3.8
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-WTF==1.0.1
Flask-Migrate==3.1.0
python-dotenv==0.21.0
gunicorn==20.1.0

# Testing
pytest==7.1.2
pytest-cov==3.0.0
email-validator==1.1.3
pytest-mock==3.7.0

# Payments
paystackapi==2.1.3

# Email
Flask-Mail==0.9.1
PyJWT==2.3.0
